drm_hwcomposer: Add basic validation when mapping layers to planes.
Currently, we map layers to planes and expect the commit to pass.
However, we can have cases where the plane cannot support the
required format or setting properties like rotation or
pre-multiplied alpha. In such cases we need to composite the
layer using GL. This patch adds support for this. We could be
clever about choosing the current or next plane depending on
size etc but that would need detail power analysis and is left
for future. We dont yet deal cases where test commit fails with
EBusy, so we can see flicker in such cases.

BUG=#11
TEST=System boots up and commit failures caused due to format
     mismatch are not present.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>