drm_hwcomposer: Add support for basic validation in drmplane.
It's important to know if a plane can handle the composition
of a given layer or do we need to fallback to GL composition.
While it's hard to check for everything without going through
Test Commit, we can handle basic things like format checks,
if rotation, alpha properties are supported etc. This patch
adds the needed support for basic validation of HWCLayer
which will be taken into use in the coming up patches to
handle Test Commit during planning phase.

BUG=#11
TEST=System boots up and commit failures caused due to format
     mismatch are not present.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>